+++ components/CompoundText.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
-import * as Types from "../types";
+import * as Types from "../common/types";
 
 interface CompoundTextProps extends Types.ComponentWithChildren {
   Name: string;
-};
+}
 
 const CompoundText: React.FC<CompoundTextProps> = (props) => {
   return <>CompoundText</>;
