+++ components/Editor.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import * as Types from "../types";
+import * as Types from "../common/types";
 
 type EditorProps = {
-    Name: string;
-    Width: string;
-    Height: string;
-    DirtyInfo?: Types.TrueFalse;
+  Name: string;
+  Width: string;
+  Height: string;
+  DirtyInfo?: Types.TrueFalse;
 };
 
 const Editor: React.FC<EditorProps> = (props: EditorProps) => {
