+++ components/MenuItem.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
-import * as Types from "../types";
+import * as Types from "../common/types";
 
 interface MenuItemProps extends Types.ComponentWithChildren {
   Name: string;
   Label: string;
   MenuItemCode?: string;
-};
+}
 
 const MenuItem: React.FC<MenuItemProps> = (props) => {
   return <>MenuItem</>;
